[MAJ] Possibilité de générer les versions alpha, beta et rc à partir du trunk. exemple : ./build-version -s a1

+++ build-version.sh
@@ -3,18 +3,24 @@
 # Functions
 contains() { [[ $1 =~ (^|[[:space:]])"$2"($|[[:space:]]) ]] && return 0 || return 1; }
 
-usage() { echo "Usage: $0 [-b <branch>] [-t]" 1>&2; exit 1; }
+usage() { echo "Usage: $0 [-b <branch>] [-s <special version for the trunk. a1 to create alpha 1 for instance>] [-t]" 1>&2; exit 1; }
 
 # Parameters
 
 branchesList="3.0 4.0 4.1"
 bflag=0
-while getopts b:ht name
+sflag=0
+tflag=0
+while getopts b:hs:t name
   do
     case $name in
         b)
             bflag=1
             bval="$OPTARG"
+            ;;
+        s)
+            sflag=1
+            sval="$OPTARG"
             ;;
         t)
             tflag=1
@@ -35,6 +41,16 @@ then Branch=$bval;
 else Branch='4.1';
 fi
 
+if [ $sflag == 1 ] ;
+then
+Branch='master'
+bflag=0
+tflag=0
+localRepositoryDir='trunk';
+else
+localRepositoryDir='phpboost-'$Branch;
+fi
+
 # Script beginning
 
 scriptDir=$(pwd)
@@ -47,14 +63,19 @@ Original='phpboost'
 versionControlUtil='git'
 remoteRepositoryUrl='https://github.com/PHPBoost/PHPBoost.git'
 localRepositoryPath='..'
-localRepositoryDir='phpboost-'$Branch
 
 if [ ! -d $localRepositoryPath/$localRepositoryDir ] ;
-then echo 'cloning repository'
-cd $localRepositoryPath;
-mkdir $localRepositoryDir;
-$versionControlUtil clone --branch $Branch $remoteRepositoryUrl $localRepositoryDir
-cd $scriptDir;
+then
+	echo 'cloning repository'
+	cd $localRepositoryPath;
+	mkdir $localRepositoryDir;
+	$versionControlUtil clone --branch $Branch $remoteRepositoryUrl $localRepositoryDir
+	cd $scriptDir;
+else
+	echo 'updating repository'
+	cd $localRepositoryPath/$localRepositoryDir;
+	$versionControlUtil pull
+	cd $scriptDir;
 fi
 
 echo 'beginning building'
@@ -66,17 +87,32 @@ ln -s $localRepositoryPath/$localRepositoryDir $Original
 mkdir -p $buildsDir
 rm -rf $Build
 
+
 ## Build version zip
+if [ $sflag == 1 ] ;
+then
+echo $sval > $Original/kernel/.build
+else
 build_version=$(cat $Original/kernel/.build)
-#build_version=$((${build_version} + 1))
-#echo $build_version > $Export'/kernel/.build'
+fi
 
-mkdir -p $exportDir/diff
+if [ $tflag == 1 ] ;
+then
+build_version=$(cat $Original/kernel/.build)
+build_version=$((${build_version} + 1))
+echo $build_version > $Original'/kernel/.build'
+# TODO : COMMIT new build_version and tag
+fi
 
-if [ $build_version > 0 ] ;
-then ./make-diff.sh phpboost-$Branch.$(($build_version - 1)) phpboost-$Branch.$build_version $Branch;
-cp export/diff/* $exportDir/diff/
-rm -rf export/diff/
+if [ $sflag != 1 ] ;
+then
+	mkdir -p $exportDir/diff
+
+	if [ $build_version > 0 ] ;
+	then ./make-diff.sh phpboost-$Branch.$(($build_version - 1)) phpboost-$Branch.$build_version $Branch;
+	cp export/diff/* $exportDir/diff/
+	rm -rf export/diff/
+	fi
 fi
 
 echo 'copying files'
@@ -84,14 +120,18 @@ cp -r $Original/ $Build
 
 ## Nettoyage des dossiers
 rm -rf $Build'/.gitignore' $Build'/.git' $Build'/.settings' $Build'/.project' $Build'/.htaccess' $Build'/test' $Build'/update' $Build'/HomePage' $Build'/server_migration.php' $Build'/todo.txt' $Build'/changelog.txt' $Build'/templates/phpboost' $Build'/README.md'
+
 ## Suppression des fichiers .empty
-find $Build -name '.empty' -exec rm -rf '{}' \;
+find $Build -name '.empty' -exec rm -f '{}' \;
 
 ## Htaccess
 touch $Build/.htaccess
 
 ## CHMOD
-chmod -R 777 $Build'/cache' $Build'/upload' $Build'/templates' $Build'/menus' $Build'/lang'  $Build'/images' $Build'/'
+chmod -R 777 $Build'/cache' $Build'/upload' $Build'/templates' $Build'/lang' $Build'/images' $Build'/'
+if [ $Branch == '3.0' ] || [ $Branch == '4.0' ] || [ $Branch == '4.1' ] ;
+then chmod -R 777 $Build'/menus'
+fi
 
 
 ################################ Full pack ######################################
@@ -117,18 +157,21 @@ rm -rf $Build_full'/install/distribution/'
 
 ################################ PDK pack ######################################
 
-echo 'building PDK pack'
-rm -rf $Build_pdk
-cp -r $Build/ $Build_pdk
-rm -rf $Build_pdk'/articles' $Build_pdk'/calendar' $Build_pdk'/contact' $Build_pdk'/online' $Build_pdk'/shoutbox' $Build_pdk'/faq' $Build_pdk'/forum' $Build_pdk'/gallery' $Build_pdk'/web' $Build_pdk'/guestbook' $Build_pdk'/ThemesSwitcher' $Build_pdk'/LangsSwitcher' $Build_pdk'/media' $Build_pdk'/news' $Build_pdk'/newsletter' $Build_pdk'/pages' $Build_pdk'/customization' $Build_pdk'/sitemap' $Build_pdk'/search' $Build_pdk'/poll' $Build_pdk'/stats' $Build_pdk'/download' $Build_pdk'/wiki'
+if [ $sflag != 1 ] ;
+then
+	echo 'building PDK pack'
+	rm -rf $Build_pdk
+	cp -r $Build/ $Build_pdk
+	rm -rf $Build_pdk'/articles' $Build_pdk'/calendar' $Build_pdk'/contact' $Build_pdk'/online' $Build_pdk'/shoutbox' $Build_pdk'/faq' $Build_pdk'/forum' $Build_pdk'/gallery' $Build_pdk'/web' $Build_pdk'/guestbook' $Build_pdk'/ThemesSwitcher' $Build_pdk'/LangsSwitcher' $Build_pdk'/media' $Build_pdk'/news' $Build_pdk'/newsletter' $Build_pdk'/pages' $Build_pdk'/customization' $Build_pdk'/sitemap' $Build_pdk'/search' $Build_pdk'/poll' $Build_pdk'/stats' $Build_pdk'/download' $Build_pdk'/wiki'
 
-rm -rf $Build_pdk'/install/distribution.ini'
-rm -rf $Build_pdk'/install/lang/french/distribution.php'
-rm -rf $Build_pdk'/install/lang/english/distribution.php'
-mv $Build_pdk'/install/distribution/distribution_pdk.ini' $Build_pdk'/install/distribution.ini'
-mv $Build_pdk'/install/distribution/distribution_pdk_french.php' $Build_pdk'/install/lang/french/distribution.php'
-mv $Build_pdk'/install/distribution/distribution_pdk_english.php' $Build_pdk'/install/lang/english/distribution.php'
-rm -rf $Build_pdk'/install/distribution/'
+	rm -rf $Build_pdk'/install/distribution.ini'
+	rm -rf $Build_pdk'/install/lang/french/distribution.php'
+	rm -rf $Build_pdk'/install/lang/english/distribution.php'
+	mv $Build_pdk'/install/distribution/distribution_pdk.ini' $Build_pdk'/install/distribution.ini'
+	mv $Build_pdk'/install/distribution/distribution_pdk_french.php' $Build_pdk'/install/lang/french/distribution.php'
+	mv $Build_pdk'/install/distribution/distribution_pdk_english.php' $Build_pdk'/install/lang/english/distribution.php'
+	rm -rf $Build_pdk'/install/distribution/'
+fi
 
 ## Documentation
 #mkdir -p $Build_pdk'/doc/'$Branch
@@ -137,19 +180,21 @@ rm -rf $Build_pdk'/install/distribution/'
 ################################
 
 ## Export modules zip
+if [ $sflag != 1 ] ;
+then
+	echo 'building modules zip files'
+	cd $Build
 
-echo 'building modules zip files'
-cd $Build
+	rm -rf $scriptDir/$exportDir/modules
+	mkdir -p $scriptDir/$exportDir/modules
 
-rm -rf $scriptDir/$exportDir/modules
-mkdir -p $scriptDir/$exportDir/modules
-
-for folder in $(ls); 
-do 
-	if [ -e $folder/config.ini ] ; 
-	then zip -r $scriptDir/$exportDir/modules/$folder.zip $folder 1>/dev/null;
-	fi
-done
+	for folder in $(ls); 
+	do 
+		if [ -e $folder/config.ini ] ; 
+		then zip -r $scriptDir/$exportDir/modules/$folder.zip $folder 1>/dev/null;
+		fi
+	done
+fi
 
 ################################
 
@@ -159,8 +204,14 @@ cd $scriptDir
 ## Export to zip
 mkdir -p $exportDir/phpboost
 cd $buildsDir/
-zip -r $scriptDir/$exportDir/phpboost/phpboost.zip phpboost/ 1>/dev/null
-zip -r $scriptDir/$exportDir/phpboost/phpboost_pdk.zip phpboost_pdk/ 1>/dev/null
+
+if [ $sflag != 1 ] ;
+then 
+	zip -r $scriptDir/$exportDir/phpboost/phpboost.zip phpboost/ 1>/dev/null
+	zip -r $scriptDir/$exportDir/phpboost/phpboost_pdk.zip phpboost_pdk/ 1>/dev/null
+else
+	zip -r $scriptDir/$exportDir/phpboost/phpboost_$sval.zip phpboost/ 1>/dev/null
+fi
 
 cd $scriptDir
 rm -rf $Build
@@ -168,6 +219,9 @@ unlink $Original
 rm -rf 0
 
 echo -e ''
-echo 'zip files are created for PHPBoost '$Branch'.'$build_version
+if [ $sflag != 1 ] ;
+then echo 'zip files are created for PHPBoost '$Branch'.'$build_version;
+else echo 'zip files are created for the new PHPBoost special release '$sval;
+fi
 
 exit 0
